services/admin-api/pyproject.toml
[project]
name = "sastaspace-admin-api"
version = "0.1.0"
description = "Admin API for sastaspace: system metrics, container status, SSE logs, STDB write proxy."
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.118",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "tenacity>=9.0",
    "psutil>=6.0",
    "docker>=7.0",
    "google-auth>=2.35",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-q --cov=src --cov-report=term-missing --cov-fail-under=70"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]
